refactor: merge bouncer and collector into probeservices
To implement https://github.com/ooni/probe/issues/887, it helps to have all the
logic for speaking with probeservices in the same place.
