+++ probeservices/bouncer.go
@@ -1,20 +1,22 @@
-// Package bouncer contains a OONI bouncer client implementation.
+// Package probeservices contains code to contact OONI probe services.
 //
 // Specifically we implement v2.0.0 of the OONI bouncer specification defined
-// in https://github.com/ooni/spec/blob/master/backends/bk-004-bouncer.md.
-package bouncer
+// in https://github.com/ooni/spec/blob/master/backends/bk-004-bouncer
+//
+// We additionally implement v2.0.0 of the OONI collector specification defined
+// in https://github.com/ooni/spec/blob/master/backends/bk-003-collector.md.
+package probeservices
 
 import (
 	"context"
-	"net/http"
 
 	"github.com/ooni/probe-engine/internal/jsonapi"
 	"github.com/ooni/probe-engine/model"
 )
 
-// Client is a client for the OONI bouncer API.
+// Client is a client for the OONI probe services API.
 type Client struct {
-	// BaseURL is the bouncer base URL.
+	// BaseURL is the probe services base URL.
 	BaseURL string
 
 	// HTTPClient is the HTTP client to use.
@@ -29,9 +31,7 @@ type Client struct {
 
 // GetCollectors queries the bouncer for collectors. Returns a list of
 // entries on success; an error on failure.
-func (c *Client) GetCollectors(
-	ctx context.Context,
-) (output []model.Service, err error) {
+func (c *Client) GetCollectors(ctx context.Context) (output []model.Service, err error) {
 	err = (&jsonapi.Client{
 		BaseURL:    c.BaseURL,
 		HTTPClient: c.HTTPClient,
@@ -43,8 +43,7 @@ func (c *Client) GetCollectors(
 
 // GetTestHelpers is like GetCollectors but for test helpers.
 func (c *Client) GetTestHelpers(
-	ctx context.Context,
-) (output map[string][]model.Service, err error) {
+	ctx context.Context) (output map[string][]model.Service, err error) {
 	err = (&jsonapi.Client{
 		BaseURL:    c.BaseURL,
 		HTTPClient: c.HTTPClient,
