+++ session.go
@@ -11,7 +11,6 @@ import (
 	"time"
 
 	"github.com/ooni/probe-engine/atomicx"
-	"github.com/ooni/probe-engine/bouncer"
 	"github.com/ooni/probe-engine/geoiplookup/iplookup"
 	"github.com/ooni/probe-engine/geoiplookup/mmdblookup"
 	"github.com/ooni/probe-engine/geoiplookup/resolverlookup"
@@ -27,6 +26,7 @@ import (
 	"github.com/ooni/probe-engine/model"
 	"github.com/ooni/probe-engine/netx/bytecounter"
 	"github.com/ooni/probe-engine/netx/httptransport"
+	"github.com/ooni/probe-engine/probeservices"
 )
 
 // SessionConfig contains the Session config
@@ -474,7 +474,7 @@ func (s *Session) maybeLookupCollectors(ctx context.Context) error {
 	if len(s.availableCollectors) > 0 {
 		return nil
 	}
-	return s.queryBouncer(ctx, func(client *bouncer.Client) (err error) {
+	return s.queryBouncer(ctx, func(client *probeservices.Client) (err error) {
 		s.availableCollectors, err = client.GetCollectors(ctx)
 		return
 	})
@@ -529,20 +529,20 @@ func (s *Session) maybeLookupTestHelpers(ctx context.Context) error {
 	if len(s.availableTestHelpers) > 0 {
 		return nil
 	}
-	return s.queryBouncer(ctx, func(client *bouncer.Client) (err error) {
+	return s.queryBouncer(ctx, func(client *probeservices.Client) (err error) {
 		s.availableTestHelpers, err = client.GetTestHelpers(ctx)
 		return
 	})
 }
 
-func (s *Session) queryBouncer(ctx context.Context, query func(*bouncer.Client) error) error {
+func (s *Session) queryBouncer(ctx context.Context, query func(*probeservices.Client) error) error {
 	s.queryBouncerCount.Add(1)
 	for _, e := range s.getAvailableBouncers() {
 		if e.Type != "https" {
 			s.logger.Debugf("session: unsupported bouncer type: %s", e.Type)
 			continue
 		}
-		err := query(&bouncer.Client{
+		err := query(&probeservices.Client{
 			BaseURL:    e.Address,
 			HTTPClient: s.DefaultHTTPClient(),
 			Logger:     s.logger,
